CRUD curso

+++ backend/controllers/usuarios.js
@@ -1,5 +1,6 @@
 const validator = require('validator');
 const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
 
 const Grupo = require('../models/grupos');
 const Usuario = require('../models/usuarios');
@@ -27,7 +28,7 @@ const getUsuarios = async(req, res) => {
             ]);
         } else {
             [usuarios, total] = await Promise.all([
-                Usuario.find({}).skip(from).limit(registropp).populate('grupo'),
+                Usuario.find({}).skip(from).limit(registropp).populate('grupo', '-__v'),
                 Usuario.countDocuments()
             ]);
         }
@@ -97,45 +98,75 @@ const actualizarUsuario = async(req, res = response) => {
 
     const { password, email, ...object } = req.body;
     const uid = req.params.id;
-    try {
 
+    const token = req.header('x-token');
+    const decodedUserId = jwt.verify(token, process.env.JWTSECRET).uid;
+
+    try {
         // Comprobamos si está intentando cambiar el email, que no coincida con alguno ya existente
         // Buscamos si ya existe un email igual que el que nos llega en el post
         const existeEmail = await Usuario.findOne({ email: email });
+        // si hay un email igual al pasado por le body
         if (existeEmail) {
-            // Comprobamos que el email sea el suyo, el UID debe ser igual, que esté en uso.
+            // comprobamos si el id del model y el id pasado por parametro son distintos
             if (existeEmail._id != uid) {
                 return res.status(400).json({
                     ok: false,
                     msg: 'Email ya existe'
                 });
             }
-            // si llegamos aquí, el email no está en uso o no está en la BD
-            object.email = email;
-            const usuario = await Usuario.findByIdAndUpdate(uid, object, { new: true });
-            res.json({
-                ok: true,
-                msg: 'Usuario actualizado'
+        }
+        // si llegamos aquí, el email no está en uso o no está en la BD
+        // pero primero comprobamos si el id es distinto, porque eso no se puede hacer
+        // TODO: hacer que los admin tambien puedan hacerlo
+        if (decodedUserId != id) {
+            return res.status(400).json({
+                ok: false,
+                msg: 'Identificador no valido'
             });
         }
+        object.email = email;
+        const usuario = await Usuario.findByIdAndUpdate(uid, object, { new: true });
+        res.json({
+            ok: true,
+            msg: 'Usuario actualizado'
+        });
     } catch (error) {
         console.log(error);
         return res.status(400).json({
             ok: false,
             msg: 'Error actualizando usuario'
-        })
+        });
     }
 
 }
 
 const borrarUsuario = async(req, res) => {
     const uid = req.params.id;
+    // recuperamos el token
+    const token = req.header('x-token');
+    // buscamos el id del token para comprobar si el usuario que solicita el borrado es el mismo id que el del parametro
+    const { id, rol } = jwt.verify(token, process.env.JWTSECRET);
+
+    // comprobamos si hay token
+    if (!token) {
+        return res.status(400).json({
+            ok: false,
+            msg: 'El Token es necesario'
+        });
+    }
     try {
         const existeUsuario = await Usuario.findById(uid);
         if (!existeUsuario) {
             return res.status(400).json({
                 ok: true,
                 msg: 'Usuario no existe'
+            });
+        }
+        if (id != uid || rol != 'ROL_ADMIN') {
+            return res.status(400).json({
+                ok: true,
+                msg: 'No tienes permiso para realizar la accion'
             });
         }
         const resultado = await Usuario.findByIdAndRemove(uid);
